# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: RankMap API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/clusters:
    get:
      operationId: listClusters
      tags: [clusters]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of clusters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KeywordCluster"
    post:
      operationId: createCluster
      tags: [clusters]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClusterBody"
      responses:
        "201":
          description: Created cluster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeywordCluster"
...
  /projects/{projectId}/clusters/{id}:
    get:
      operationId: getCluster
      tags: [clusters]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cluster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeywordCluster"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCluster
      tags: [clusters]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClusterBody"
      responses:
        "200":
          description: Updated cluster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeywordCluster"
    delete:
      operationId: deleteCluster
      tags: [clusters]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/clusters/{id}/approve:
    post:
      operationId: approveCluster
      tags: [clusters]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved cluster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeywordCluster"
  /projects/{projectId}/clusters/{id}/reject:
    post:
      operationId: rejectCluster
      tags: [clusters]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected cluster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeywordCluster"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    KeywordCluster:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        tenantId:
          type: integer
        label:
          type: string
        pillarTopic:
          type: ["string", "null"]
        clusterType:
          type: string
          enum: [pillar, cluster, supporting]
        status:
          type: string
          enum: [pending, approved, rejected]
        keywordCount:
          type: integer
        avgScore:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - tenantId
        - label
        - clusterType
        - status
        - keywordCount
        - createdAt
        - updatedAt
    CreateClusterBody:
      type: object
      properties:
        label:
          type: string
        pillarTopic:
          type: string
        clusterType:
          type: string
          enum: [pillar, cluster, supporting]
      required:
        - label
    UpdateClusterBody:
      type: object
      properties:
        label:
          type: string
        pillarTopic:
          type: string
        clusterType:
          type: string
          enum: [pillar, cluster, supporting]
...